requirements.txt
certifi==2018.4.16
chardet==3.0.4
cloudpickle==0.5.3
cycler==0.10.0
Cython==0.28.5
dask==0.18.2
decorator==4.3.0
Django==2.1
django-bootstrap-form==3.4
django-filter==2.0.0
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.2.0
idna==2.7
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
networkx==2.1
numpy==1.15.0
opencv-contrib-python==3.4.2.17
Pillow==5.2.0
PyJWT==1.6.4
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
requests==2.19.1
scikit-image==0.14.0
scipy==1.1.0
six==1.11.0
toolz==0.9.0
urllib3==1.23
